+++ lib/utils/consts.dart
@@ -3,6 +3,7 @@ const enableNotificationsFlag = 'enable_notifications';
 const recordLocationFlag = 'record_location';
 const enableMatrixFlag = 'enable_matrix';
 const enableTooltipFlag = 'enable_tooltip';
+const enableAiStreamingFlag = 'enable_ai_streaming';
 const resendAttachments = 'resend_attachments';
 const enableLoggingFlag = 'enable_logging';